requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0
psycopg2-binary==2.9.9
pydantic==2.7.3
requests==2.32.3
passlib[bcrypt]
python-jose[cryptography]
python-dotenv==1.0.1
alembic==1.13.1